% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dmask.R
\name{calculate_d_mask}
\alias{calculate_d_mask}
\title{calculates the D subset (rank orderings mask), and number of levels and unique values per margin}
\usage{
calculate_d_mask(dat, mode)
}
\arguments{
\item{dat}{data matrix.}

\item{mode}{Margin type, must be one of "fixed", "multi" or "mixed".}
}
\value{
List containing the D mask, number of unique values per margin, and unique margin values
}
\description{
calculates the D subset (rank orderings mask), and number of levels and unique values per margin
}
\keyword{internal}
