% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_loglik.R
\name{calculate_log_lik}
\alias{calculate_log_lik}
\title{calculate log-likelihood by integrating over the posterior distribution}
\usage{
calculate_log_lik(k_lambda, k_omega, alpha, dat)
}
\arguments{
\item{k_lambda}{stored lambda matrix from mcmc iteration.}

\item{k_omega}{stored omega matrix from mcmc iteration.}

\item{alpha}{stored alpha vector from mcmc iteration.}

\item{dat}{data matrix.}
}
\value{
log-likelihood
}
\description{
calculate log-likelihood by integrating over the posterior distribution
}
\keyword{internal}
