% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmBPFA_sampler.R
\name{chain_call}
\alias{chain_call}
\title{Meta-function that bundles the sampler into a chain call}
\usage{
chain_call(
  pre_sparse,
  warmup,
  iter,
  thin,
  dat,
  mode,
  share_noise,
  d_mask,
  n_levels,
  margin_vals,
  x,
  lambda,
  zeros,
  omega,
  alpha,
  gamma_k,
  d,
  dc,
  ibp_a,
  ibp_b,
  benchmark
)
}
\description{
Performs single chain run of the mmBPFA algorithm.
}
\keyword{internal}
