% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{check_accuracy}
\alias{check_accuracy}
\title{Function to check reconstruction accuracy}
\usage{
check_accuracy(dat, results, true_vals = NULL, mode, confusion = T)
}
\arguments{
\item{dat}{Data matrix.}

\item{results}{processed results from \code{process_results}. Must not include standard errors.}

\item{true_vals}{Optional list of true latent values from \code{generate_sim_data}.}

\item{mode}{Margin type. Must be one of "fixed", "multi" or "mixed".}

\item{confusion}{Boolean indicating whether to return a confusion matrix. Only works for "fixed" margin data.}
}
\value{
List of observed accuracy and accuracy metrics (rmse and mae). If confusion is TRUE, also returns confusion matrix results.
}
\description{
\code{check_accuracy} ingests the cleaned results from \code{process_results} without standard errors and 
returns the accuracy in reconstructing the observed data. If \code{true_vals} from \code{generate_sim_data} 
is supplied, the function will also return the RMSE and MAE of posterior predictions for the latent values. 
A confusion matrix, precision, recall, and Kappa score are also available via \code{confusion = TRUE} for data
with fixed marginal distributions.
}
