% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_k_start}
\alias{check_k_start}
\title{Selects appropriate K_start based on rank of data}
\usage{
check_k_start(K_start, p, n)
}
\description{
Selects appropriate K_start based on rank of data
}
\keyword{internal}
