% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{destructure.mcmc.output.processed}
\alias{destructure.mcmc.output.processed}
\title{zeallot destructure function}
\usage{
\method{destructure}{mcmc.output.processed}(x)
}
\description{
zeallot destructure function
}
\keyword{internal}
