% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{destructure.mmBPFA.simulated.data}
\alias{destructure.mmBPFA.simulated.data}
\title{zeallot destructure function}
\usage{
\method{destructure}{mmBPFA.simulated.data}(x)
}
\description{
zeallot destructure function
}
\keyword{internal}
