% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{extract_omega_lambda_mean_se}
\alias{extract_omega_lambda_mean_se}
\title{function to estimate factor means and se's}
\usage{
extract_omega_lambda_mean_se(
  x,
  dim_order,
  stable_K,
  p,
  se = TRUE,
  to_tibble = TRUE
)
}
\description{
function to estimate factor means and se's
}
\keyword{internal}
