% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_sim_data}
\alias{generate_sim_data}
\title{Generate data for running simulations with mmBPFA}
\usage{
generate_sim_data(
  mode = "fixed",
  n_levels = 2,
  n = 1000,
  p = 500,
  K_true = 10,
  sparsity = NULL,
  a = 10,
  b = 1,
  d = 2,
  e = 1,
  prop_missing = 0,
  noise = 0.05,
  seed = 123
)
}
\arguments{
\item{mode}{Sets the type of margins for the simulated data. Must be one of "fixed", "multi" or "mixed." Default is "fixed."}

\item{n_levels}{Max number of unique values per margin. When mode is "fixed" n_levels is constant across margins. When mode is "multi" a margin may take on unique values up to n_levels. When mode is "mixed", n_levels determines max number of arbitrary distributions data may be drawn from. Default is 2.}

\item{n}{Number of observations. Default is 1000.}

\item{p}{Number of features. Default is 500.}

\item{K_true}{True number of latent dimensions. Default is 10.}

\item{sparsity}{Numeric value between 0 and 1 that determines amount of sparsity in the factor loadings. If sparsity argument is set, zeros are uniformly distributed across all dimensions. Default is NULL.}

\item{a}{A hyperparameter determining the distribution of sparsity across the factor loadings. Default is 10}

\item{b}{A hyperparameter determining the distribution of sparsity across the factor loadings. Default is 1.}

\item{d}{A hyperparameter determining the shape of the factor precisions (gamma_k). Default is 2.}

\item{e}{A hyperparameter determining the scale of the factor precisions (gamma_k). Default is 1.}

\item{prop_missing}{Numeric value between 0 and 1 determining the proportion of missingness in the data. Default is 0.}

\item{noise}{Numeric value between 0 and 1 determining how much noise should be induced into the data. Default is 0.05.}

\item{seed}{Random seed. Default is 123.}
}
\value{
S3 object of class \code{mmBPFA.simulated.data} containing generated data and true underlying values. Attributes note the user-given arguments.
}
\description{
\code{generate_sim_data} produces simulated data that can be used for testing
the mmBPFA sampler prior to production runs to ensure it will behave as expected.
}
