% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dims_2d}
\alias{get_dims_2d}
\title{function to get extract dimensions of 2d objects}
\usage{
get_dims_2d(x, warn = FALSE)
}
\description{
function to get extract dimensions of 2d objects
}
\keyword{internal}
