% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_init.R
\name{initialize_sampler}
\alias{initialize_sampler}
\title{generate initial values for single chain}
\usage{
initialize_sampler(dat, mode, n_levels, margin_vals, K_start = 100)
}
\arguments{
\item{dat}{Data matrix.}

\item{mode}{Margin type.}

\item{n_levels}{Numeric vector of number of unique values per margin.}

\item{margin_vals}{Numeric vector or list of unique values per margin.}

\item{K_start}{Number of latent dimensions to intialize sampler with. Default is 100.}
}
\description{
generate initial values for single chain
}
\keyword{internal}
