\name{mmBPFA-package}
\alias{mmBPFA-package}
\alias{mmBPFA}
\docType{package}
\title{\packageTitle{mmBPFA}}
\description{\packageDescription{mmBPFA}}
\section{Package Content}{\packageIndices{mmBPFA}}
\author{\packageAuthor{mmBPFA}}
\section{Maintainer}{\packageMaintainer{mmBPFA}}
\keyword{package}
