% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_benchmarks}
\alias{process_benchmarks}
\title{Function to process benchmarks}
\usage{
process_benchmarks(mcmc_object)
}
\arguments{
\item{mcmc_object}{A S3 object of class \code{bpfa.results} containing benchmarks.}
}
\value{
A list containing the raw benchmarks and average timing per sample step.
}
\description{
\code{process_benchmarks} ingests the raw results from \code{mmBPFA_sampler} and
returns the raw benchmarks and average time per sampling step. Only works if the sampler was run with
\code{benchmarks = TRUE}.
}
