% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Process results from mmBPFA sampling run}
\usage{
process_results(mcmc_object, SE = T, p = c(0.05, 0.95), to_tibble = T)
}
\arguments{
\item{mcmc_object}{An S3 object of class \code{bpfa.results} from mmBPFA sampling run.}

\item{SE}{A boolean indicating whether to calculate standard error credibility intervals around posterior estimates.}

\item{p}{a vector indicating the lower and upper probability bounds for the credibility intervals. Default is c(0.05, 0.95).}

\item{to_tibble}{A boolean indicating whether to convert output to tibbles for use in tidy workflow. Default is TRUE.}
}
\value{
An S3 object of class \code{mcmc.output.processed} containing the processed posterior margins.
}
\description{
\code{process_results} ingests raw results from \code{mmBPFA_sampler} and returns processed estimates of posterior quantities of interest.
}
