% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_inactive.R
\name{remove_inactive}
\alias{remove_inactive}
\title{Remove inactive features}
\usage{
remove_inactive(zeros, lambda, omega, gamma_k, dc)
}
\arguments{
\item{zeros}{P x K sparsity-inducing binary matrix.}

\item{lambda}{P x K matrix of factor loadings.}

\item{omega}{N x K matrix of factor scores.}

\item{gamma_k}{A numeric vector factor precisions of length K.}

\item{dc}{A numeric vector counting number of times each dimension has been sampled.}
}
\value{
List of zeros, lambda, omega, gamma_k, dc.
}
\description{
Remove inactive features
}
\keyword{internal}
