% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_alpha}
\alias{sample_alpha}
\title{Sampling step for alpha (item-intercepts)}
\usage{
sample_alpha(x, lambda, omega)
}
\arguments{
\item{x}{A multivariate Gaussian copula matrix of dimenion N x P.}

\item{lambda}{A standard multivariate normal matrix of factor loadings of dimension P x K.}

\item{omega}{A standard multivariate normal matrix of factor scores of dimenion N x K.}
}
\value{
alpha Numeric vector of item-intercepts of length P.
}
\description{
Sampling step for alpha (item-intercepts)
}
\keyword{internal}
