% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_d}
\alias{sample_d}
\title{Sampling step for d (hyperparameter for sample_gamma_k)}
\usage{
sample_d(zeros, gamma_k, c, c0, d0)
}
\arguments{
\item{zeros}{A binary, sparsity inducing matrix of dimenion P x K.}

\item{gamma_k}{A gamma-distributed vector of factor precisions of length K.}

\item{c}{A tunable hyperparameter for the shape parameter of the gamma draw. Default is 1.}

\item{c0}{A tunable hyperparameter for the shape parameter of the gamma draw.}

\item{d0}{A tunable hyperparameter for the inverse-scale (rate) parameter of the gamma draw.}
}
\value{
scalar double
}
\description{
Sampling step for d (hyperparameter for sample_gamma_k)
}
\keyword{internal}
