% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_gamma_k}
\alias{sample_gamma_k}
\title{Sampling step for gamma_k (factor precisions)}
\usage{
sample_gamma_k(zeros, lambda, c = 0, d = 100)
}
\arguments{
\item{zeros}{A binary, sparsity inducing matrix of dimenion P x K.}

\item{lambda}{A standard multivariate normal matrix of factor loadings of dimension P x K.}

\item{c}{A tunable hyperparameter for the shape parameter of the gamma draw. Default is 0.}

\item{d}{A tunable hyperparameter for the inverse-scale (rate) parameter of the gamma draw. Default is 100}
}
\value{
gamma_k Numeric vector of factor precisions of length K.
}
\description{
Sampling step for gamma_k (factor precisions)
}
\keyword{internal}
