% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_IBP_a}
\alias{sample_IBP_a}
\title{Sampling step for Indian Buffet Process Alpha Parameter}
\usage{
sample_IBP_a(ibp_b, zeros, e = 1, f = 1)
}
\arguments{
\item{ibp_b}{A scalar double, Indian Buffet Process beta parameter.}

\item{zeros}{A binary, sparsity inducing matrix of dimenion P x K.}

\item{e}{A scalar double, tunable hyperparameter for the shape parameter of the gamma draw. Default is 1.}

\item{f}{A scalar double, tunable hyperparameter for the rate parameter of the gamma draw. Default is 1.}
}
\value{
ibp_a Scalar double, Indian Buffet Process alpha parameter.
}
\description{
Sampling step for Indian Buffet Process Alpha Parameter
}
\keyword{internal}
