% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_IBP_b}
\alias{sample_IBP_b}
\title{Metropolis-Hastings sampling step for Indian Buffet Process Beta Parameter}
\usage{
sample_IBP_b(ibp_a, ibp_b, zeros)
}
\arguments{
\item{ibp_a}{A scalar double, Indian Buffet Process alpha parameter.}

\item{ibp_b}{A scalar double, prior Indian Buffet Process Beta parameter .}

\item{zeros}{A binary, sparsity inducing matrix of dimenion P x K.}
}
\value{
ibp_b Scalar double, Indian Buffet Process beta parameter.
}
\description{
Metropolis-Hastings sampling step for Indian Buffet Process Beta Parameter
}
\keyword{internal}
