% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_lambda_and_zeros}
\alias{sample_lambda_and_zeros}
\title{Sampling step for lambda and binary, sparsity-inducing zero matrix}
\usage{
sample_lambda_and_zeros(
  x,
  lambda,
  zeros,
  omega,
  alpha,
  gamma_k,
  dc,
  ibp_a,
  ibp_b,
  tau,
  sparse,
  infinite
)
}
\arguments{
\item{x}{A multivariate Gaussian copula matrix of dimenion N x P.}

\item{lambda}{A standard multivariate normal matrix of factor loadings of dimension P x K.}

\item{zeros}{A binary, sparsity inducing matrix of dimenion P x K.}

\item{omega}{A standard multivariate normal matrix of factor scores of dimenion N x K.}

\item{alpha}{A numeric vector of item-intercepts of length P.}

\item{gamma_k}{A numeric vector of factor precisions of length K.}

\item{dc}{A integer vector of counts for the number of times each dimensions has been sampled.}

\item{ibp_a}{A scalar double, Indian Buffet Process beta parameter.}

\item{ibp_b}{A scalar double, Indian Buffet Process beta parameter.}

\item{tau}{A tunable hyperparameter for the poisson draws of new dishes to be sampled.}

\item{sparse}{A boolean indicating whether to include sparsity-inducing prior.}

\item{infinite}{A boolen indicating whether to sample new potential dishes at each step.}
}
\value{
zeros A binary, sparsity inducing matrix of dimenion P x K.

lambda A standard multivariate normal matrix of factor loadings of dimension P x K.

dc A integer vector of counts for the number of times each dimensions has been sampled.
}
\description{
Sampling step for lambda and binary, sparsity-inducing zero matrix
}
\keyword{internal}
