% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_omega}
\alias{sample_omega}
\title{Sampling step for omega matrix (Factor Scores) of dimension N x K.}
\usage{
sample_omega(lambda, x, alpha)
}
\arguments{
\item{lambda}{A standard multivariate normal matrix of factor loadings of dimension P x K.}

\item{x}{A multivariate Gaussian copula matrix of dimenion N x P.}

\item{alpha}{A numeric vector of item-intercepts of length P.}
}
\value{
omega A multivariate Gaussian matrix of factor scores of dimension N x K.
}
\description{
Sampling step for omega matrix (Factor Scores) of dimension N x K.
}
\keyword{internal}
