% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_x.R
\name{sample_x}
\alias{sample_x}
\title{sample X values according to Hoff (2007) and Murray et al. (2013) extended rank likelihood approach}
\usage{
sample_x(dat, mode, d_mask, n_levels, margin_vals, x, lambda, alpha, omega)
}
\arguments{
\item{dat}{Data matrix.}

\item{mode}{Margin type. Must be one of "fixed", "mixed" or "multi."}

\item{d_mask}{Boolean mask of extended rank likelihood.}

\item{n_levels}{Numeric vector of number of unique values per margin.}

\item{margin_vals}{List of unique values per margin.}

\item{x}{N x P Gaussian copula matrix.}

\item{lambda}{P x K matrix of factor loadings.}

\item{alpha}{A numeric vector of item-level intercepts.}

\item{omega}{N x K matrix of factor scores.}
}
\value{
List of x matrix and cutpoints.
}
\description{
sample X values according to Hoff (2007) and Murray et al. (2013) extended rank likelihood approach
}
\keyword{internal}
