% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmBPFA_sampler.R
\name{sampler_phase}
\alias{sampler_phase}
\title{All-in-one function for the sampling procedure}
\usage{
sampler_phase(
  phase,
  dat,
  mode,
  share_noise,
  d_mask,
  n_levels,
  margin_vals,
  x,
  lambda,
  alpha,
  omega,
  zeros,
  gamma_k,
  d,
  dc,
  ibp_a,
  ibp_b,
  benchmark
)
}
\description{
All-in-one function for the sampling procedure
}
\keyword{internal}
