% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vectorized_which.max}
\alias{vectorized_which.max}
\title{vectorized which.max function for D mask}
\usage{
vectorized_which.max(x, y)
}
\description{
vectorized which.max function for D mask
}
\keyword{internal}
